requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJWT
cryptography
celery 
redis 
django-celery-beat
gevent
python-dotenv
django-celery-results
django-cors-headers
django-redis
django-ratelimit
dj-database-url
pycryptodome
psycopg2
psycopg2-binary
python-json-logger
google-generativeai
google-api-core
# OCR Dependencies - Updated for Python 3.13
pytesseract==0.3.13
Pillow>=10.0.0
opencv-python>=4.10.0.84
numpy>=1.26.0  # Compatible with Python 3.13

# Image Processing
opencv-python-headless==4.10.0.84  # Alternative if GUI not needed

# Utilities
python-magic==0.4.27
python-dateutil==2.9.0
pytz==2024.1

# Code Quality
black==24.8.0
flake8==7.1.1
isort==5.13.2

python-magic-bin

django-storages[boto3]
pytest
pytest-django
pytest-cov 
pytest-xdist 
factory-boy 
faker 
freezegun 
pytest-mock 
responses


